+++ app/javascript/components/Login.vue
@@ -59,8 +59,6 @@
           username: this.credentials.username,
           password: this.credentials.password
         }
-        // We need to pass the component's this context
-        // to properly make use of http in the auth service
         auth.login(this, credentials, '/balance')
       }
     }